misc - fix code style in picture_cleanup.class.php

+++ htdocs/util2/cron/modules/picture_cleanup.class.php
@@ -1,14 +1,13 @@
 <?php
 /***************************************************************************
- *  For license information see doc/license.txt
+ * For license information see doc/license.txt *
  *
- *
- *  Delete duplicate log pictures (produced e.g. by Ocprop)
+ * Delete duplicate log pictures (produced e.g. by Ocprop)
  ***************************************************************************/
 
-checkJob(new picture_cleanup());
+checkJob(new PictureCleanup());
 
-class picture_cleanup
+class PictureCleanup
 {
     public $name = 'picture_cleanup';
     public $interval = 86400;
@@ -25,7 +24,7 @@ class picture_cleanup
 
         while ($rDuplicatePic = sql_fetch_assoc($rsDuplicatePic)) {
             $rsInstances = sql(
-                " SELECT `pictures`.`id` `picid`, `cache_logs`.`cache_id` `cache_id`
+                "SELECT `pictures`.`id` `picid`, `cache_logs`.`cache_id` `cache_id`
                  FROM `pictures`
                  LEFT JOIN `cache_logs` ON `cache_logs`.`id` = `pictures`.`object_id`
                  WHERE `pictures`.`object_type`=1 AND `pictures`.`object_id`='&1' AND `pictures`.`title`='&2'
@@ -44,20 +43,19 @@ class picture_cleanup
                 if ($instances[$n]['filesize'] !== false) {
                     // ensure that pic is stored locally
                     for ($nn = $n - 1; $nn >= 0; -- $nn) {
-                        if ($instances[$nn]['filesize'] === $instances[$n]['filesize']) {
-                            if (file_get_contents($instances[$nn]['pic']->getFilename())
-                                == file_get_contents($instances[$n]['pic']->getFilename())
-                            ) {
-                                $picture = $instances[$n]['pic'];
-                                echo
-                                    'deleting duplicate picture '
-                                    . $picture->getPictureId() . ' ("' . $picture->getTitle() . '")'
-                                    . ' from log ' . $rDuplicatePic['object_id']
-                                    . ' of cache ' . $instances[$n]['cache_id'] . "\n";
-                                $picture->delete(false);
-                                $instances[$n]['filesize'] = false;
-                                break;
-                            }
+                        if ($instances[$nn]['filesize'] === $instances[$n]['filesize']
+                            && file_get_contents($instances[$nn]['pic']->getFilename())
+                                === file_get_contents($instances[$n]['pic']->getFilename())
+                        ) {
+                            $picture = $instances[$n]['pic'];
+                            echo
+                                'deleting duplicate picture '
+                                . $picture->getPictureId() . ' ("' . $picture->getTitle() . '")'
+                                . ' from log ' . $rDuplicatePic['object_id']
+                                . ' of cache ' . $instances[$n]['cache_id'] . "\n";
+                            $picture->delete(false);
+                            $instances[$n]['filesize'] = false;
+                            break;
                         }
                     }
                 }
